Draw bus during simulation

+++ src/main/java/domain/IDistributableElement.java
@@ -1,5 +1,5 @@
 package domain;
 
 public interface IDistributableElement {
-    public float generate();
+    public double generate();
 }
